driver/requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.1
werkzeug==3.0.3
zipp==3.19.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.1
werkzeug==3.0.3
zipp==3.19.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.1
werkzeug==3.0.3
zipp==3.19.1
